requirements.txt
pandas
torch
transformers
utils
torch
transformers
prepro
scikit-learn  
simcse
matplotlib
tiktoken
apted
openai
tree_sitter_languages
google-generativeai